- Fix failing 'DocumentsServiceTest' setting upload date.

Conflicts:
	src/main/java/de/unistuttgart/vis/vita/model/document/Document.java
	src/main/java/de/unistuttgart/vis/vita/model/entity/EntityRelation.java

+++ src/test/java/de/unistuttgart/vis/vita/analysis/importer/epub/Epub2ImportModuleTest.java
@@ -63,7 +63,7 @@ public class Epub2ImportModuleTest {
   @Test
   public void testMetadataPublishYear() {
     int text = 1993;
-    assertEquals(text, epub2Result.getMetadata().getPublishYear());
+    assertEquals(text, epub2Result.getMetadata().getPublishYear().intValue());
   }
 
   @Test
